fix: minimax.cpp

+++ backtracking/minimax.cpp
@@ -39,7 +39,7 @@ int minimax(int depth, int node_index, bool is_max,
  */
 int main() {
     std::array<int, 8> scores = {90, 23, 6, 33, 21, 65, 123, 34423};
-    int height = log2(scores.size());
+    double height = log2(scores.size());
 
     std::cout << "Optimal value: " << backtracking::minimax(0, 0, true, scores, height)
               << std::endl;
